feat(server): convert to human-friendly URLs
Offers performance as the sacrificial lamb to the tech debt gods in exchange for human-friendly URLs.
Builds and runs are indexed by project IDs and build IDs, so they cannot be fetched until the build ID and project ID are resolved now that they are not in the URL. This would be solved by a graphQL-like single-shot query, but we're not there yet.

fixes #40

BREAKING CHANGE: Server URLs are now of the form `/app/projects/:projectSlug/compare/:partialBuildId` replacing `/app/projects/:projectId/builds/:buildId`