Trim down Dockerfile
We noticed that the image size could be greatly reduced by using
a smaller base image. This migrates the base image to node:12-buster-slim,
which provides the same functionality but shrinks the image size from
1.5GB to 720MB.
Also, the apt-cache gets cleaned up after package installation to avoid keeping
it in a layer of the Docker image.

This also updates the image to the latest Node version 12.