Add new fixture folder
To really test more edge cases I added a new folder with 6 pages, this
way we can test the default limit of 5 as well disabling the limit.

First I tried to do it in the same fixture folder but that obviously
broke other tests because they were pickng up the wrong pages.